requirements.txt
aiohttp
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
bitarray==2.8.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.1
cytoolz==0.12.2
Django==4.2.3
djangorestframework==3.14.0
eth-abi==4.2.1
eth-account==0.9.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.5.1
eth-utils==2.3.0
frozenlist==1.4.0
gunicorn==20.1.0
hexbytes==0.3.1
idna==3.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
lru-dict==1.2.0
multidict==6.0.4
packaging==23.2
parsimonious==0.9.0
protobuf==4.24.4
psycopg2==2.9.9
py-solc-x==2.0.2
pycparser==2.21
pycryptodome==3.19.0
pyOpenSSL==23.2.0
pytz==2023.3
pyunormalize==15.0.0
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
rlp==3.0.0
rpds-py==0.10.6
sqlparse==0.4.4
toolz==0.12.0
typing_extensions==4.7.1
urllib3==2.0.3
web3==6.11.1
websockets==12.0
yarl==1.9.2
